games/firstCPPGame: FInish setting up Physics
Changes:
* Vector2 and Point2 are now made out of floats and not ints.
+ New clampPoint2.
* Fixed use of limiting borders with new Physics.

Functionality is now the same as in commit 1828b17, but with more options to play around with the movement.

+++ FirstCPPGame/FirstCPPGame/render.cpp
@@ -16,14 +16,14 @@ namespace Render {
 	void
 		drawRectInPixels(Physics::Vector2 rect, Utils::u32 color) {
 		rect = Physics::clampVector2(
-			{ {0, 0}, {RENDER_STATE.WND_WIDTH, RENDER_STATE.WND_HEIGHT} }, 
+			{ {0, 0}, {(float)RENDER_STATE.WND_WIDTH, (float)RENDER_STATE.WND_HEIGHT} },
 			rect
 		);
 
 		for (int y = rect.start.y; y < rect.end.y; y++) {
 			Utils::u32* pixel = 
 				(Utils::u32*)RENDER_STATE.WND_MEMORY + 
-				rect.start.x +
+				(int)rect.start.x +
 				y * 
 				RENDER_STATE.WND_WIDTH;
